#1507 Заставить работать и для LongPoll
